Initial commit
This has been a private project that I have tinkered with and evolved over a period
of years. Much of the C++ Hearts-specific code is several years old, and
is related to work that I did a couple decades ago while experimenting with an
algorithm to play Hearts using Decision Trees.

All of the python and TensorFlow code is recent, but has evolved organically over
a period of several months. During that time I did a lot of experimentation, and
did not use good story/branch discipline. Primarily because of that, this first
public release of the project is starting fresh with all prior history of development
removed.

Even though I am starting fresh with this release, there is still a fair amount of
"dirty laundry", i.e. the code is not as clean as I would normally develop when
working in a team that uses good practices of pull requests with code reviews. It
is my intention to use better discipline going forward.