Remove dropout layers in CNN
Testing showed that the dropout layers were impeding accuracy.
In hindsight it does seem reasonable that dropout could only hurt.
We don't have to worry about overfitting since we are easily able
to generate a lot of data and hold out half of it for validation.
We are want to be somewhat stingy with the number of units in the CNN,
and for dropout to really add value it is probably necessary to be
generous with the number of units, to get the ensembler of classifiers
effect.