Consolidate DnnMonteCarlo and SimpleMonteCarlo
This is a significant refactor. SimpleMonteCarlo can now be used
for either of the two use cases. We simply pass in a different "intuition"
strategy.

Along the way, I also cleaned up the use of annotators, and used
smart pointers for both Annotator and Strategy objects.

I intend to rename SimpleMonteCarlo -> MonteCarlo, but I'll do
that in a separate commit.

Also, I think MonteCarlo needs to be a little smarter than simply being
told how many alternates to play. The amount of time to play out a game
from the first trick is much longer than the amount time to play out the game
from the last few tricks. We can't expect much precision from our estimates
early int the game anyway, so it is probably fine to use fewer alternates
at first and gradually increase as we go.

Ultimately we want something that primarily uses a time budget.