Run symmetric threads from the outer loop
After all of the previous work for pooled predictor and
being able to run multiple threads from MonteCarlo::choosePlay,
this commit instead implements symmetric threads from the
outer loop of the hearts generator app, which yields good performance.

I still plan to use the multiple threads from choosePlay()
so that in real game play more alternates can be explored in
the same amount of time, but this commit doesn't exercise that
capability.