Simplify training loop with better stopping condition
This story started out with the intent to implement dlib's
[Automatic Learning Rate Scheduling](http://blog.dlib.net/2018/02/automatic-learning-rate-scheduling-that.html)
but in experimentation I found that Adam gives better results.

However, various related changes are nice. The training loop
is simpler and more efficient when running with a GPU.

The dlib ALRS is used only to provide another stopping condition,
though it will probably need to be tweaked over time.