Bug fixes and improvements
This work meandered over a period of a month of intermittent
testing and tweaking. I found some bugs that may have corrupted
some of the generated data in subtle ways, so it is might be
best to regenerate the first round of data and build new models.
However, I was still able to demonstrate that DnnModelIntuition
player using a generated model is very nearly as strong as
a MonteCarlo player using random intuition. Better still,
a MonteCarlo player using the model intuition with just a few rollouts
is stronger than player used to generate the data that built
the model.