Revise types used for Rank, Suit
This is in anticipation of making the new code in
websockets/CardsWebServer.[ch]pp use the types in lib/Card.h.

But rather that simply modifying CardsWebServer.hpp, I wanted to
see if the somewhat better type safety in CardsWebServer.hpp
could be ported over to the main project. It turns out it wasn't
hard to do.